requirements.txt
numpy
PyQt6
pyqtgraph
PyYAML
librosa
tensorflow-cpu
soundfile